LPS-111561 Isolate Pagination selector
